\name{fromtoby}
\alias{fromtoby}
\title{Create Regular Sequence from 'from-to-by' Vector}
\description{
  This function creates a sequence from named vectors with the 
  names \code{from}, \code{to} and \code{by}.
}
\usage{
fromtoby(times)
}
\arguments{
  \item{times}{A named vector with the names  \code{from}, \code{to} and
    \code{by}.
  }
}
\details{
  Named vectors with  \code{from}, \code{to} and \code{by} can be used
  in \pkg{simecol} to specify time steps.
}
\value{
  The appropriate vector with a sequence, generated by \code{seq}.
}
\seealso{\code{\link[base]{seq}}}
\examples{
times <- c(from=1, to=5, by=0.1)
fromtoby(times)
}
\keyword{manip}

