\name{modelFit-class}
\docType{class}
\alias{modelFit-class}

\title{Class of Fitted Model Parameters}
\description{
  Class that contains parameters and other information returned by \code{\link{fitOdeModel}}.
}
  

\section{Methods}{
  \code{\link{coef}}, \code{\link{deviance}}, \code{\link{print}}
}

\seealso{\code{\link{fitOdeModel}}}

\keyword{classes}

