\name{listOrNULL-class}
\docType{class}
\alias{listOrNULL-class}
\alias{functionOrNULL-class}
\alias{numericOrlist-class}

\title{Helpful Union Classes}
\description{Classes representing either lists or NULL (i.e. empty),
  functions or NULL, or numeric vector or list respectively.}
\section{Objects from the Class}{These classes are virtual: No objects may
  be created from it.}
\section{Methods}{
No methods exist for these classes.
}

\seealso{\code{\link{simObj}}}

\keyword{classes}

