\name{listOrNULL-class}
\docType{class}
\alias{listOrNULL-class}
\alias{listOrdata.frame-class}
\alias{functionOrNULL-class}
\alias{numericOrlist-class}
\alias{functionOrcharacter-class}

\title{Helpful Union Classes}
\description{
  Classes representing either list or NULL (i.e. empty), function or
  NULL, function or character vector, numeric vector or list, or list or
  data.frame.
}
  
\section{Objects from the Class}{
  These classes are virtual: No objects may
  be created from it.
}

\section{Methods}{
  No methods exist for these classes.
}

\seealso{\code{\link{simObj}}}

\keyword{classes}

